// apps/desktop/src/types/generated/TargetRefDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DefKeyDto } from "./DefKeyDto";

/**
 * Mirrors [`TargetRef`].
 */
export type TargetRefDto = { 
/**
 * The [`FieldRole::TargetKey`](rim_resolve::domain::FieldRole::TargetKey) field's canonical path text.
 */
keyField: string, 
/**
 * The target def itself.
 */
def: DefKeyDto, };